refactor: Merges struct types
Throughout the API Gateway and DynamoDB code, there were similarities
that were suitable to merge into common struct types.

The benefit of this is that bespoke "conversion" logic isn't necessary
if HandInfo is a 1:1 to its DynamoDB counterpart. So in theory the
body coming in from the request can immediately be used in the `PutItem`
API call.

This results in (hopefully) maintable code, while also still being
terse.
