+++ backend/src/aws/dynamodb.go
@@ -9,21 +9,9 @@ import (
 	"github.com/aws/aws-sdk-go/aws"
 )
 
-type DDBHandler struct {
-	DynamoDbClient *dynamodb.Client
-	TableName      string
-}
-
-type Hand struct {
-	PlayerName string  `dynamodbav:"playerName"`
-	PlayerId   string  `dynamodbav:"playerId"`
-	Version    string  `dynamodbav:"version"`
-	Cards      []int16 `dynamodbav:"cards"`
-}
-
 // TODO: Handle Duplicates - this should return an error
 // TODO: Need to check that types match before marshalling
-func (ddbh DDBHandler) AddHand(h Hand) error {
+func (ddbh DDBHandler) AddHand(h HandInfo) error {
 	item, err := attributevalue.MarshalMap(h)
 	if err != nil {
 		log.Panicf("unable to marshal submitted hand: %v", err)
